batchspawner.py: Don't use exec_prefix when running connect_to_job_cmd
This is a breaking change, and probably needs to be made optional. It is
needed when exec_prefix of the form "ssh <remotehost>" to launch commands
remotely, because the same prefix will not be appropriate for the "connect
to job" command, which will need to add the port forwarding to the ssh
command.